refactor(mobile): split goal derive
